backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# OpenAI integration
openai==1.3.8

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Database (optional)
sqlalchemy==2.0.23
alembic==1.12.1

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
aiofiles==23.2.1
Pillow==10.1.0 